requirements.txt
# Dependencias mínimas para Streamlit Cloud - Dashboard Enterprise
streamlit>=1.49.1
plotly>=6.3.0
pandas>=2.3.2
numpy>=2.3.3
openpyxl>=3.1.0
python-dateutil>=2.8.0
PyYAML>=6.0